DI-4623 Fix to latest posts not limiting to set number
Using the 'latest posts' block, it does not seem to limit the number of posts to display even if a number is set.

Occurring both if block on home page or another page. Issue does not seem related to number of columns.

For example if I have 5 posts and in the settings for the latest post block I have limited in settings to show maximum 3 posts over 3 columns in visual preview mode the correct (3) posts show however when saved on the live page all 5 posts show over 3 columns/2 rows.